Started building code to create a table
I worked with my tutor and we decided to start to build a table because there wasn't salary data that would allow us to make a gauge using salary.  I am going to use the company data to identify how many times a company has an open for the job title that is searched.